added tabs to new movie fragment + added info to tabs
