+++ app/src/main/java/ru/surf/course/movierecommendations/adapters/GridTVShowsAdapter.java
@@ -12,7 +12,7 @@ import java.util.List;
 import ru.surf.course.movierecommendations.GridViewHolder;
 import ru.surf.course.movierecommendations.MainActivity;
 import ru.surf.course.movierecommendations.R;
-import ru.surf.course.movierecommendations.fragments.MovieInfoFragment;
+import ru.surf.course.movierecommendations.fragments.OldMovieInfoFragment;
 import ru.surf.course.movierecommendations.models.MovieInfo;
 import ru.surf.course.movierecommendations.models.TVShowInfo;
 import ru.surf.course.movierecommendations.tmdbTasks.ImageLoader;
@@ -53,8 +53,8 @@ public class GridTVShowsAdapter extends RecyclerView.Adapter<GridViewHolder> {
     }
 
     private void fragmentToSwitch(MovieInfo info) {
-        MovieInfoFragment movieInfoFragment = MovieInfoFragment.newInstance(info);
-        switchContent(R.id.activity_main_container, movieInfoFragment);
+        OldMovieInfoFragment oldMovieInfoFragment = OldMovieInfoFragment.newInstance(info);
+        switchContent(R.id.activity_main_container, oldMovieInfoFragment);
     }
 
     public void setTvShowInfoList(List<TVShowInfo> list) {
