+++ app/src/main/java/ru/surf/course/movierecommendations/adapters/ListMoviesAdapter.java
@@ -15,7 +15,7 @@ import java.util.List;
 
 import ru.surf.course.movierecommendations.MainActivity;
 import ru.surf.course.movierecommendations.R;
-import ru.surf.course.movierecommendations.fragments.MovieInfoFragment;
+import ru.surf.course.movierecommendations.fragments.OldMovieInfoFragment;
 import ru.surf.course.movierecommendations.models.MovieInfo;
 import ru.surf.course.movierecommendations.tmdbTasks.ImageLoader;
 
@@ -59,8 +59,8 @@ public class ListMoviesAdapter extends RecyclerView.Adapter<ListMoviesAdapter.My
     }
 
     private void fragmentToSwitch(MovieInfo info) {
-        MovieInfoFragment movieInfoFragment = MovieInfoFragment.newInstance(info);
-        switchContent(R.id.activity_main_container, movieInfoFragment);
+        OldMovieInfoFragment oldMovieInfoFragment = OldMovieInfoFragment.newInstance(info);
+        switchContent(R.id.activity_main_container, oldMovieInfoFragment);
     }
 
     public void setMovieInfoList(List<MovieInfo> list) {
